// src/app/Life/services/quotation.interface.ts
export interface IClientDetails {
  clientNo: string;
  fullName: string;
  phoneNumber: string;
  website?: any;
  email: string;
  title: string;
  maritalStatus: string;
  firstName: string;
  middleName: string;
  surname: string;
  salutation?: any;
  clientType: string;
  gender: string;
  address: string;
  kyc?: any;
  crmId: string;
  idNumber?: any;
  alternativeEmail: string;
  alternativePhoneNumber: string;
  dateCreated: Date;
  dateOfBirth: Date;
  category: string;
  providerNo: string;
  redgNo?: any;
  noOfEmployee?: any;
  noOfGroups?: any;
  financialSizeCategory?: string;
  sector: string;
  providerCategory: string;
  providerCode?: any;
  providerSubCategory: string;
  state?: any;
  band: string;
  employer?: any;
  language: string;
  bankNo: string;
  enroleeNo: string;
  occupationGroup?: any;
  socialMedia?: any;
  clv: string;
  communication?: any;
}

export interface ICurrency {
  description: string;
  validCurrency: string;
}
export interface IDiscountCode {
  description: string;
  code: string;
}
export interface IPremiumFrequency {
    id:          number;
    premFreq:    string;
    description: string;
    noPayYear:   number;
    createdBy:   string;
    frequency:   string;
}
export interface IPremiumPaymentMethod {
  description: string;
  code: string;
}

export interface ICodes {
  codeSubgroup: string,
  code: string,
  codeGroup: string,
  title: string,
  codeCat: string,
  active: boolean,
  recType: string,
  id: number,
  deleted: any
}

export interface IAssetDetails {
  assetNo: number;
  registrationNo: string;
}

export interface IAccessoryType {
  codeSubgroup: string;
  code: string;
  codeGroup: string;
  title: string;
  codeCat?: any;
  active: boolean;
  recType: string;
  id: number;
  deleted?: any;
}

export interface IBenefitPaymentFreq {
  paymentFrequency: string;
}

export interface ITargetContributionFreq {
  frequency: string;
}
export interface IProduct {
  description: string;
  code: string;
}

export interface IContentCat {
  codeSubgroup: string;
  code: string;
  codeGroup: string;
  title: string;
  codeCat: string;
  active: boolean;
  recType: string;
  id: number;
  deleted: string;
}

export interface IInsuranceType {
  title: string;
  code: string;
}

export interface IAssuredRelToOwner {
  relationship: string;
}

export interface ICoverInfo {
    id:                                 number;
    productCode:                        string;
    subgroupNo:                         string;
    availablePostIssue:                 boolean;
    sumAssuredEqualBaseCoverSumAssured: boolean;
    blankSumAssured:                    boolean;
    mandatory:                          boolean;
    company:                            string;
    ageAssuredLimit:                    string;
    coverCode:                          string;
    coolOffPeriod:                      number;
    fundingOption:                      string;
    maxNo:                              string;
    description:                        string;
    base:                               string;
    code:                               string;
    coverSumAssured:                    string;
    riderCoverCode:                     string;
}
export enum CoverScreenNames {
    BL = "BUILDING",
    DR = "DEFENSE_RECOURSE", 
    LIL = "LIFE_INSURANCE_LIFE", 
    AL = "ACCESSORIES_RISK", // /
    TP = "THIRDPARTY", 
    PC = "PROPERTY_CONTENT", 
    SAV = "SAVINGS", 
    MH = "MARINE_HULL", 
    MC = "MARINE_CARGO",
    LA = "LIFE_ANNUITY",
    AC = "ANNUITIES_CERTAIN",
    ILR = "INDIVIDUAL_LIFE_RISK",
    ICL = "INDIVIDUAL_CREDIT_LIFE",
    IS = "INDIVIDUAL_SAVINGS",// /
    END = "ENDOWMENT",
    GLR = "GROUP_CREDIT_LIFE",// /
    BO = "BOND"
}

export type CoverCode = 'AC' | 'LA' | 'ILR' | 'ICR'| 'IS' | 'END' | 'GLR' | 'GS' | 'TP' | 'DR' | 'AL' | 'LIL' | 'BL' | 'PC' | 'SAV' | 'MH' | 'MC' | 'ICL' | 'BO'

export interface ICoverScreen {
  description: string;
  coverScreen: CoverCode;
}

export interface ICoversDetails {
  info: ICoverInfo;
  screen: {
    code: CoverCode;
    name: CoverScreenNames;
  };
}